requirements.txt
diffusers
transformers
accelerate
opencv-python
numpy
Pillow
torch
scipy
safetensors
gradio
